add -Dfile.encoding=UTF-8 JVM Option
this is follow-up measures with f96444b9657008057be1a394e0b3e1598851b4fd

details:
the f96444b works fine with macOS because the default value of
file.encoding is UTF-8 on macOS.
but the Windows default file.encoding is MS949. so it makes wrong
interactions with *real* Actlist application.